[DUT-833]: 온보딩 step 컴포넌트 분리 및 전환 로직 정리
