+++ src/pages/OnboardingWardCreatePage/types.ts
@@ -1,0 +1,1 @@
+export type TSortMode = 'manual' | 'employmentDate';
